requirements.txt
asgiref==3.5.0
cycler==0.11.0
Django==4.0.4
djangorestframework==3.13.1
Faker==10.0.0
fonttools==4.29.1
httplib2==0.20.4
kiwisolver==1.3.2
magicmock==0.3
matplotlib==3.5.1
networkx==2.7
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
